move conftest up
